Version bump and adds the applications key to make Firefox happy.
The version bump will bump both RECAP and Firefox extensions, since it's higher than the current value for either. The 'applications' key is suggested by Firefox's linter to ensure that our calls to storage work properly. Without this, we get a temporary ID, which is OK, but not ideal, and besides, without this we won't be able to upgrade the current extension.

More info: https://developer.mozilla.org/en-US/Add-ons/WebExtensions/WebExtensions_and_the_Add-on_ID