Adds fall back if the referrer didn't work.
This fixes a number of bugs that were filed in Github, though, as indicated by the comment I added, I don't know why we need to use the referrer info at all. I'm leaving it for now, but I'm hopeful to remove it or understand it better in the future.