Restore httpRequest() for binary data.
That is, remove jQuery AJAX for binary files, and use XmlHTTPRequest
via the home-grown helper function.

This (partially) reverts commit 9e1d3b3760af15ac88e401779b25c4432df91e93.

Bascally, it doesn't work reliably, but it is extremely
system-dependent and not entirely clear why.

I get files that are stuffed with 0xefbfbd strings that balloon the
PDF files and corrupt htem, but Steve and Mike don't have this
problem. But I see it in both Chrome and Firefox.

For instance:

> 0000000: 2550 4446 2d31 2e36 0a25 efbf bdef bfbd  %PDF-1.6.%......
> 0000010: efbf bdef bfbd 0a32 2030 206f 626a 0a3c  .......2 0 obj.<

when it should be:

> 0000000: 2550 4446 2d31 2e36 0a25 e2e3 cfd3 0a32  %PDF-1.6.%.....2
> 0000010: 2030 206f 626a 0a3c 3c2f 4c65 6e67 7468   0 obj.<</Length

In that case, it pushes a 106285 PDF into a 183815.
Problem occurs in both Firefox and Chrome.

Also, pass the xhr to the callback to ease in debugging.