Fixes freelawproject/recap#182.
The main fix here was to bring back httpRequest function, which John did. From there, I simplified the number of object conversions we did by removing the arrayBufferToArray function and the conversions that it did. Uploads were broken due to https://github.com/freelawproject/courtlistener/commit/83edd7312527dcbe5346e3ce8e55e99feeab4974, and that's fixed too. I think we're back in business, but we'll see if it works for everybody.