content_delegate: "Waiting for download..." message
Although this grew out of other debugging, it seems worth saving.

When you click "view," a page is thrown up with am empty iframe that
is later populated with the PDF blob when the download completes.
DIsplay the words "Waiting for download..." on that page prior to
download completion. Generally this won't be visible, but if it is,
it's a better user experience than a blank page.