Support external PDF viewers via option
This is most of a fix for
https://github.com/freelawproject/recap/issues/183 however it does not
quite achieve the ideal of following the browser-configured PDF
choice. On the other hand, recap-firefox didn't do that either -- it
always forced external PDF downloading as a consequence of support for
setting a filename. In a perfect world, we would like to better, but
this achieves fixes our regression and more importantly stops the
suffering for those who don't want in-browser PDF viewing.

* Add an `external_pdf` option, which defaults to false (background.js)
* Manifest it as checkbox (options.html)
* Actually do it (content_delegate.js):
. Rename updateHtmlPage() to displayPDF() to better define what it does
. Reword explanatory comment -- it's hard to decide how much of the
  comments should explain deferred actions that run as a result of
  callbacks.
. Only createObjectURL(blob) if we're not doing external viewing
. Call out to FileSaver.js's saveAs() method to do the work, plus
  remove the "Waiting for dowanload..." message.
. Linebreak between complex callback definition and execution.
. This code needs to be re-indented. Separate commit to follow.