pacer.js: Relax /doc1 url regex constraint
doc1 urls come in many shapes and sizes. They can be as simple as

https://ecf.cand.uscourts.gov/doc1/035116111385

or at least as complex as:

https://ecf.cand.uscourts.gov/doc1/035116111385?caseid=316722&de_seq_num=1129&magic_num=

Don't require them to match /doc1/\d+$, that is all digits at the end.
Instead, match /doc1/\d+ without anchoring to the end. So we won't
pick up non-digits in the first match, but we will function on the
slightly more complex doc1 links (2nd example above) avaialble from,
e.g., NEFs.