findPacerDocIds: do current page
Store the current page's docid/caseid mapping, not only the mappings
from links that are embedded in this page.

xxx: Why does "this.url" evaluate to "undefined" but "this.links" seems to work?
Is this maybe a discrepancy between the debugger and reality? It seems like the
closure should be sufficient. I don't quite get what's going on here.