content_delgate: parse doDocPostURL() xxx
Parse doDocPostURL() onclick calls to get case ID.

There is some schizophrenia between the internal case ID and the
docket number that we will subsequently have to resolve. But we are
using the docket number reported by CMECF as the RECAP case ID for the
time being, consistent with historical practice and because that may
be the only viable choice, given the information we have available.

Start to parse goDPU(), but leave it out for now given caseID issues.
xxx Should use array destructuring.