FileSaver.min.js -> FileSaver.js
Replace minified FileSaver with un-minified from
https://raw.githubusercontent.com/eligrey/FileSaver.js/c347c51f6e6e56221b2fa3d8c9b32e181f76355c/FileSaver.js

Mike suggests that un-minified code is a better bet for extensions,
given the liklyhood of future debugging.

Certainly we're not concerned with download speed in an extensino, nor
with obfuscataion.

And arguably if minification of FileSaver was worth it, it'd be worth
it to minify the entire extension, and we could do it at that time.

(On the other hand, debugging of FileSaver is a lot less likely than
debugging of the rest of the extension.)