Use storePacerDocIdsMultipleCases abstraction
Don't store docid/caseid mappings directly in chrome.storage without
indirecting through a function in recap.js. Previously that was
storePacerDocIds(), but that only takes a single caseid, and we need multiple.

So define storePacerDocIdsMultipleCases(), and use it.
Also make storePacerDocIds() use it too.

Why?

Because in the future we may (likely will) revisit the decision to
store key/value pairs in the toplevel of the chrome storage. Because
what if a document id conflicts with some other kind of thing we want
to store there? Or we want to store other sorts of name/value pairs.

Abstraction for the win.