fix(case number parser): Prefer the current URL over the referrer
The current code prefers parsing results of the referrer for a page over the url
for the page itself. That doesn't make sense. This change flips the logic so we
prefer parsing results of the page we're on over the results of the page we were
on moments ago.

Discussion of this at: https://github.com/freelawproject/recap-chrome/commit/2a8d5a8a2c9df62a13f4a259cca0b07408a23f9b