fix(case number parser): Guard against caseId params
In appellate PACER, URLS can have a caseId parameter, but it means something different
than the caseid parameter in district and bankruptcy PACER. Typically, it has a value
of caseId=0, so we just guard against that. If we get that value, we press on.

Discussion of this at: https://github.com/freelawproject/recap-chrome/commit/d8754eb910c47f8298d9df1fea387ab74a702268#commitcomment-26188087