pacer.js: Adjust a99f656 comments, semicolon
Log message in a99f656 was incorrect. Clarification:

Appellate CMECF does not use caseId for somethign different -- it is
the same thing.  But when Appellate refers to a District CMECF docket,
it doesn't know the caseid, and rather than leaving it unset, it sets
it explicitly to zero. Here we ignore that zero, as long as it's a
single digit.

Appellate uses "caseId" where District uses "caseid" (lowercase), so
match the regexp case-insensitively, with /i.

Missing semicolon on "continue" statement.
Javascript autosemicolonification is antidisestablishmentarian.