fix(recap.js): Don't send the pacer_case_id if null or undefined
Previously this was getting stringified, so if it was null or undefined, it became 'null' or
'undefined'. The backend would accept this, and we'd have bad results. Now, if we don't have the
pacer_case_id, we don't send it, and typically, the backend will reject it.

Fixes freelawproject/recap#224