fix(Uploads): Ensures we don't POST 'undefined' and 'null'  values
In freelawproject/recap#226, we identified that we were often sending the string "undefined" or
"null" for the pacer_case_id field in attachment page uploads. While this didn't have any
immediate effects on the server, it's not good. This change ensures that if we don't have those
values, we don't send them. In some cases, this will mean that the server will reject uploads,
but that is safer and better than sending bad data.

Previously, the server would also reject uploads that lacked a pacer_case_id value, but that has
been relaxed to only apply to dockets. Attachment pages and documents can survive without this
value being sent.

Fixes freelawproject/recap#226.