fix(parsing): Adds a new function to get the case ID from various forms
As part of fixing freelawproject/recap#226, I noticed that we weren't getting the pacer_case_id
for attachment pages. This change mirrors the recent changes we made with the pacer_doc_id field,
which made is to that we prefer to get document IDs from the goDLS function, not from the URLs.
With this change, we do something similar for the pacer_case_id. We try to get it from the inputs
and the forms, and if that fails, we fall back to the old technique of getting the value from the
URL or the referrer.

The net of this is that we'll get the pacer_case_id value more consistently and accurately.