fix(page identification): Fixes abstraction problem
As we slowly upgraded our regexes to work more consistently, we discovered an abstraction
problem between isSingleDocumentPage and isDocumentUrl, both of which used the same regexes.
This commit makes the code more DRY by making isSingleDocumentPage use isDocumentUrl.
Functionally, this should be the same, but it allows us only a single place to upgrade the
regular expressions.