fix(page identification): Better support for show_doc att pages
Attachment pages can be identified by either doc1 or show_doc URLs. Currently, we
do not identify these pages correctly if they are behind a show_doc URL. This means
that we do not upload them. With the better abstraction introduced in the prior
commit, it's easy to do this correctly, ensuring that these get uploaded when they
should.