content_delegate.js: Narrow scope of auth warning
This revises the warning Ka-Ping added in 2013 (02bcc53) to be
accurate.

The "courts" do indeed provide document authentication systems. At
bottom, of course, one can redownload the document in PACER, for a
fee, or otherwise obtain it from the Court.

But perhaps more relevantly, CM/ECF does offer a cryptographic hash
service, but it is not useful because: it is restricted by ACL to
filing users only (and is not availble to PACER users who just try the
URL); the hash procedure is not publicly documented; and, at least
superficially, it appears to require a key that is not public
available.

So instead we say there is "no effective document authentication system,"
not "no document authentication system."