pacer.js:parseGoDLSFunction(): Allow null input
Don't assume we are called with a string parameter, we might be called
with null (or undefined), in which case we should fail instead of
throwing a TypeError when trying to call null.prototype.match() [or
whatever].

The easy answer is to switch from String.prototype.match() to
RegExp.prototype.exec().

The other choice would be to guard all calls, but that's more work for
no apparent gain.