Guard parseGoDLSFunction() returns.
Normally PACER.parseGoDLSFunction() returns an Object with specific
properties, but if it fails to parse the string parameter passed to
it, then it will fail. And on failure it returns false.

So ensure all callers are prepared to deal with that falsity.

(An alternative would have been to return an empty Object on failure,
but that is harder to test for and would generate errors trying to
dereference nonexistent properties. Or to return an Object with all
properties populated but with null or empty or zero values. That seems
bad for error checking.)