goDLS() checks (#40)
* pacer.js:parseGoDLSFunction(): Allow null input

Don't assume we are called with a string parameter, we might be called
with null (or undefined), in which case we should fail instead of
throwing a TypeError when trying to call null.prototype.match() [or
whatever].

The easy answer is to switch from String.prototype.match() to
RegExp.prototype.exec().

The other choice would be to guard all calls, but that's more work for
no apparent gain.

* Guard parseGoDLSFunction() returns.

Normally PACER.parseGoDLSFunction() returns an Object with specific
properties, but if it fails to parse the string parameter passed to
it, then it will fail. And on failure it returns false.

So ensure all callers are prepared to deal with that falsity.

(An alternative would have been to return an empty Object on failure,
but that is harder to test for and would generate errors trying to
dereference nonexistent properties. Or to return an Object with all
properties populated but with null or empty or zero values. That seems
bad for error checking.)

* Additional parseGoDLSFunction() tests

For cases where it should fail and return false:
with an invalid string (not parseable), or with a non-string input
(null or undefined).

* PACER.parseGoDLSFunction(): return null on failure, not false

Better JS style.