fix(PDF Uploading): Give up gracefully on document types we don't support.
This came up when a user tried to view a bankruptcy claim document. These have a different receipt pages, which our regex failed to match. That regex is triggered by an event listener on the "View Document" so when that failed, it caused the button to fail and the download to be impossible. Not good.

The fix I put in here is to just make it so that we don't get claim documents. I don't think we want them anyway, at least not for now. There may be other types of documents that we don't know about and this should guard against those too.

Closes https://github.com/freelawproject/recap/issues/234