fix(PDF Uploading): Disables (attempted) PDF uploading on Claims pages
Further addressing freelawproject/recap#234. Claims in bankruptcy cases
contain the word "Claim" instead of the word "Image". We don't yet (?)
support bankruptcy claims, so this code just identifies such pages as
not a document page and short circuits them. This way, none of our code
is run on those pages, and the behave as if RECAP is not installed.