fix(pdf uploading): Adds more safety checks
Adds one more safety check to our document and attachment number regex to to ensure that if it
fails, the user is not blocked from downloading. Without this safety check, if the regex failed
the extension would crash and the user wouldn't be able to download the item they wanted.