Fixes claim doc1 URLs  (#41)
* tweak(PDF Uploading): Cleans up numerous references to the same form el

* fix(PDF Uploading): Give up gracefully on document types we don't support.

This came up when a user tried to view a bankruptcy claim document. These have a different receipt pages, which our regex failed to match. That regex is triggered by an event listener on the "View Document" so when that failed, it caused the button to fail and the download to be impossible. Not good.

The fix I put in here is to just make it so that we don't get claim documents. I don't think we want them anyway, at least not for now. There may be other types of documents that we don't know about and this should guard against those too.

Closes https://github.com/freelawproject/recap/issues/234

* fix(PDF Uploading): Reverts tweaks from e19206fe86198fabbe28b6a65de0de67472a1229

These tweaks were a mistake, though they didn't seem to affect functionality. It really is
our intention to have two different forms here that we manipulate. forms[0] is the original
one on the page, and forms[1] is one that we make (and monkey patch) in
handleSingleDocumentPageView.

* fix(PDF Uploading): Yanks code that blocked "Claims" docs from uploading.

I think this is better placed upstream of this function so that the code
here never gets run.

* fix(PDF Uploading): Disables (attempted) PDF uploading on Claims pages

Further addressing freelawproject/recap#234. Claims in bankruptcy cases
contain the word "Claim" instead of the word "Image". We don't yet (?)
support bankruptcy claims, so this code just identifies such pages as
not a document page and short circuits them. This way, none of our code
is run on those pages, and the behave as if RECAP is not installed.

* fix(PDF Uploading): Fixes tests.

Gotta have a table with an "Image" text node in the document now.

* fix(pdf uploading): Adds more safety checks

Adds one more safety check to our document and attachment number regex to to ensure that if it
fails, the user is not blocked from downloading. Without this safety check, if the regex failed
the extension would crash and the user wouldn't be able to download the item they wanted.