fix(security): Adds a dataType parameter to our jQuery requests
As you can see in the comment, this is important if the CourtListener
server ever gets hacked. If that were to happen, and if we didn't have
this parameter in there, the CourtListener server could be used to send
JSONP data to our users. The CourtListener APIs are only called from
PACER websites, so I think the attack surface is pretty limited, but
I'm not entirely sure of that.

Whatever the scope, this is a good precaution to take and it *might* be
the one thing necessary to make RECAP available for Opera. We'll see if
they accept it yet.