feat(pacer.js): Makes getCourtFromUrl more lenient
Before now, it required a trailing slash. That was fine, but the
postMessage event.origin value lacks the trailing slash, and so doesn't
match properly in that case. With this change, the trailing slash is
optional, but because this regex is used for security purposes, we do
not match URLs like:

https://ecf.pamd.uscourts.evildoer.com/

Includes tests to prevent regressions.