refactor(recap): Moves one function to use a listener as an example
BACKGROUND
~~~~~~~~~~
This is a step in the direction of nuking importInstance and
exportInstance, and towards doing things the "normal" way, via message
passing and listening.

As mentioned in freelawproject/recap#237, the way we're doing this now
makes it difficult to use certain functions, because those functions
can only be used in background scripts that are not *also* content
scripts. Since I need one of those functions to support zip upload in
freelawproject/recap#117, I'm working to get this fixed.

The only alternative to this that I could think of was to create a new
file called recap_background.js, which would be a background-only script
and then put my code in there, but I didn't love that approach since it
separated things arbitrarily.

IMPLEMENTATION
~~~~~~~~~~~~~~
The way I've implemented this is to move the uploadDocument function
out of Recap(), and to put it under a listener at the bottom of the
file. This listener checks for a package name of "recap", and then uses
an if statement to listen for a function mame of "uploadDocument". The
idea here is that I will move *all* of the existing functions in the
Recap() object into this listener, and it will use the package check and
function name check to run the right code.

I've done a bit of research on this approach, and it appears that
there's no right way to do this. This approach feels fairly clean to me
though, so I think I'm comfortable going with it. Other approaches
I considered were:

 - Using one listener per function. I like how this would allow us to
   have a cleaner separation of the functions (they're not just
   separated by if statements, but I didn't like how this would mean a
   lot of repetition of the package/function filtering within the
   beginning of each listener.

 - Using a simpler approach to function/package filtering. Could the
   function name just be "recap.uploadDocument" for example instead of
   using to keys in the object that we pass to the listener?

 - A few other things that aren't coming to mind right now (it's late).

BIG QUESTIONS
~~~~~~~~~~~~~
1. Can I refactor all of the functions to this format without breaking
   things? This is a big refactor, and thus dangerous. Going to need
   test coverage.

2. Is there a way we can simplify sendMessage? We went from 4 lines and
   one function call to 12 lines and one function call. That's not
   great.

3. What about tests? How do we pull this apart so that it's easy to
   also refactor the tests. This seems painful.