RecapSpec: Relax tests to allow debugging
If DEBUG=true is set in recap.js, test should not fail.

Previously, tests of recap endpoints expected uploadDocket() &c. to return a result with debug false.
So if debug is true, tests fail. Boo.

Instead, don't require exact equality in the actualData.
Merely require it to be a superset of the `expected` data, with
jasmine.objectContaining().