feat(recap): Sends the installed version to any URL in the manifest. (#48)
* feat(recap): Sends the installed version to any URL in the manifest.

Fixes: freelawproject/recap#215.

* fix(recap): Refactor to restrict sites that get install message

Doing it this way lets us leave the old match rules alone, and create
new match rules specifically for receiving the message that RECAP is
installed.

* style(recap): Reverts errant whitespace change

* fix(recap): Removes local servers from match list

These are only needed for debugging, and we probably shouldn't activate
the extension on every developer's dev server ever. Better to play nice.

* Removes spurious newline.

* fix(recap): Encloses currentVersion in a block; adds message_name.

* feat(recap): Changes the run_at param

I had this wrong, assuming that document_end was when the document was
ready. Nope. Tweaked.