fix(dependencies): Fixes security issue with sub-dependency
Github alerted us to this, and I made an effort to fix it. The dep that
is causing an issue is hoek, but it seems to be gone from
package-lock.json after upgrading karma to the latest point version.

I guess that's good. I documented the process on Twitter (why not?):

https://twitter.com/mlissner/status/1001899038898188290