pacer.js: isDocketDisplayUrl() improve RE documentation
Better document the appellate regular expression in isDocketDisplayUrl(),
because there was some confusion about whether it operated properly.
JS has no (?x) flag, so we just use a long comment.

xxx: There are three problems with the regexp that become a little
more apparent to me while doing so:

1: It should anchor to "/servlet" rather than just "servlet"
2: ?servlet need not always be the first parameter
3: it shoudl allow /TransportRoom/?... in addition to /TransportRoom?...

But this is not a functional change, just a doc fix.