checkRestrictions(): Display disabled icon next to annotation
I accept criticism on the css, because css is not my thing.

We are now centering with a pair of <div>s that are text-align: center
and display: inline-block, and then left/right aligning with another
pair of divs that are display: table and display: table-cell, with
padding. There is likely a better way to do this?

Also, perhaps some of this should make it into the stylesheet, but it
is not yet apparent what?