feat(popup): Do a fundraising pop up for all our users
Unfortunately due to issue #281, this is going to pop up the tab a
second time for some users, but it's either that or we let the CSS
bug live in the wild for a few days until the roll out is complete.

Right now some of our users have the new version and others don't.