fix(dockets): Get the pacer_case_id before inserting the link
We have several ways we try to get the pacer case ID, and prior to this
we had several variables we stored it in. Ensure that we get the ID before
we insert our alert link, and make sure that we use the same variable
everywhere so the code is less confusing.