fix(links from appellate): Makes isDocketQueryUrl more generic
Before now, the function only match URLs that has a query string with only digits. That was fine, but the
link from appellate pacer to the origin case has a different query string that includes the caseNumber parameter, and so doesn't
match properly in that case. With this change, the regex expression will match both cases.