appellate/utils.js: Use URLSearchParams() properly
Don't feed the URLSearchParams() constructor an entire URL, it takes only a query string.
Feeding it a full URL causes it to misparse the first query string, if any, merging it with the URL.
In practice this was not an actual problem, but could have been.

Similarly, don't set a URL search parameter by concatenating
"&caseid=xyz"; this fails when there is no preexisting search
string:(i.e. no ?foobar), and causes the RECAP extension to break
URLs. Instead, use the URL+URLSearchParamers abstractions.

This has the side effect of converting relative URLs to absolute.  I
don't think this is an actual problem?

Lastly, don't parse a query string by hand with split('?') -- use the URL constructor to do it.